requirements.txt
pyuac==0.15.2
watchdog==2.1.3
requests==2.26.0
Pillow==8.3.1
tkinter
pystray==0.15.1
ttkbootstrap==0.5.2
